+++ src/components/Shared/molecules/InputWithIcon/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import RoundedInput from '../../atoms/RoundedInput/index';
+import RoundedInput from "../../atoms/RoundedInput/index";
 
 import {
   InputGroup,
@@ -8,7 +8,6 @@ import {
   InputProps,
   InputGroupProps,
 } from "@chakra-ui/core";
-
 
 export type InputWithIconProps = {
   iconName: string;
