+++ src/pages/PetProfile/index.tsx
@@ -132,7 +132,13 @@ const PetProfilePage = () => {
   } as Pet);
 
   const theme = useTheme() as PetCodeTheme;
-  const breakpoint = useBreakpointValue({ base: 0, sm: 1, md: 2, lg: 3, xl: 4 }) as number;
+  const breakpoint = useBreakpointValue({
+    base: 0,
+    sm: 1,
+    md: 2,
+    lg: 3,
+    xl: 4,
+  }) as number;
 
   return (
     <Layout
